Move files and updates paths
